requirements.txt
python-documentcloud==3.1.3
requests
git+https://github.com/openai/whisper.git
tqdm
patool
bs4
